# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediFlow AI Enterprise Healthcare Platform API
  description: |
    Comprehensive RESTful API for MediFlow AI — an intelligent hospital channeling, 
    clinical decision support, e-prescription, and inventory intelligence ecosystem.
    
    ### Architecture Overview
    - **Frontend:** React 19 + TypeScript + Vite + Tailwind CSS
    - **Backend API:** ASP.NET Core 8 Web API + Entity Framework Core + PostgreSQL
    - **AI Microservice:** Python 3.11/3.12 + FastAPI + LangChain / Google Gemini
    - **Real-Time Gateway:** SignalR WebSockets for Queue & Consultation Telemetry
  version: 1.0.0
  contact:
    name: MediFlow Engineering Team
    url: https://github.com/Dilshan-Pasindu/MediFlow-AI-
servers:
  - url: http://localhost:5224
    description: Local Development Backend Server
  - url: http://localhost:8000
    description: AI Agent Microservice Engine
  - url: https://api.mediflow.ai
    description: Production Cluster Gateway
...
paths:
...
  /api/doctors:
    get:
      tags:
        - Appointments & Channeling
      summary: Search and filter channeling specialists
      description: Retrieves active doctors with specialty, availability, consultation fee, and rating stats.
      operationId: getDoctors
      parameters:
        - name: specialty
          in: query
          required: false
          schema:
            type: string
          description: Filter by medical specialty (e.g., Cardiology, Dermatology)
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Keyword search on doctor name or qualification
      responses:
        '200':
          description: Paginated doctor directory list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorSummaryDto'
  /api/doctors/{id}:
    get:
      tags:
        - Appointments & Channeling
      summary: Get full doctor profile
      description: Returns doctor bio, education, available channeling slots, and patient review breakdown.
      operationId: getDoctorById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detailed doctor profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorDetailDto'
        '404':
          description: Doctor not found.
  /api/appointments:
    post:
      tags:
        - Appointments & Channeling
      summary: Book consultation slot
      security:
        - BearerAuth: []
      description: Patient books a verified consultation slot. Assigns queue token and calculates fee.
      operationId: bookAppointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookAppointmentRequest'
      responses:
        '200':
          description: Appointment booked in Pending state awaiting payment confirmation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentResponse'
        '400':
          description: Slot unavailable, doctor inactive, or appointment in the past.
        '401':
          description: Unauthorized.

    get:
      tags:
        - Appointments & Channeling
      summary: Query appointments
      security:
        - BearerAuth: []
      description: Returns filtered appointments based on caller role (patient history or clinic schedule).
      operationId: queryAppointments
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [Pending, PaymentSubmitted, Confirmed, InConsultation, Completed, Cancelled]
      responses:
        '200':
          description: List of appointment records.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Input your JWT bearer token generated from `/api/auth/login`.
  schemas:
...
    DoctorSummaryDto:
      type: object
      properties:
        id:
          type: integer
          example: 12
        fullName:
          type: string
          example: "Dr. Gregory House"
        specialty:
          type: string
          example: "Infectious Disease"
        consultationFee:
          type: number
          example: 75.00
        rating:
          type: number
          example: 4.9
        totalReviews:
          type: integer
          example: 128
        isActive:
          type: boolean
          example: true
    DoctorDetailDto:
      allOf:
        - $ref: '#/components/schemas/DoctorSummaryDto'
        - type: object
          properties:
            bio:
              type: string
              example: "Board-certified specialist with 15+ years experience in complex diagnostics."
            qualifications:
              type: string
              example: "MBBS, MD, FACP"
            hospitalClinic:
              type: string
              example: "MediFlow Central Hospital, Suite 402"
            languages:
              type: string
              example: "English, Sinhala, Tamil"
    BookAppointmentRequest:
      type: object
      required: [doctorId, dateTime]
      properties:
        doctorId:
          type: integer
          example: 12
        dateTime:
          type: string
          format: date-time
          example: "2026-10-15T09:30:00Z"
        notes:
          type: string
          example: "Persistent dry cough and shortness of breath upon exertion."
    AppointmentResponse:
      type: object
      properties:
        id:
          type: integer
          example: 501
        appointmentNumber:
          type: string
          example: "APP-2026-0501"
        doctorId:
          type: integer
          example: 12
        doctorName:
          type: string
          example: "Dr. Gregory House"
        dateTime:
          type: string
          format: date-time
        status:
          type: string
          example: "Confirmed"
        fee:
          type: number
          example: 75.00
        queueToken:
          type: integer
          example: 4
...